Correct errors in smoothScroll function
Syntax errors that were causing the smooth scroll function to not work
properly were corrected and the appropriate offset added. The offset
was necessary to keep the navbar from covering up the header.

Note that the smoothScroll function will still not scroll all the way
down to later sections in the document i.e. portfolio and contact due
to the current length of the document.