Change styling of portfolio partial
The portfolio partial is now fully responsive for smaller screen sizes.
The flexbox for projects now reflows for screens of 600px or smaller
and changes to flex-direction column. Additionally, the max-width has
been changed to use 100% with this media-width. Thumbnails for each
project have also been made responsive to prevent overflow.