Fix issue with navbar not closing
A new function was created which simulates the clicking of the navbar
toggler when one of the links is clicked. This allows the navbar to
close and free up more screen real estate. This function and
smoothScroll run simulatenously upon click.

This function had to be created separately from smoothScroll due to the
navbar name and logo which are above the collapse are targeted in the
smoothScroll but should not be targeted by the navbar toggler.
Including that element caused the navbar to toggle open.