Add initial functionality of project modal dialog
The initial structure for the modal dialog for each project has been
created. When the user clicks on one of the project thumbnails, a
full screen modal dialog opens with information about the project and
a link to visit the full project.

The HTML elements for the modal dialog are intentionally left blank.
They are populated with Javascript functionality depending on which
project is clicked.