Add modal dialog functions to script file
Javascript functionality has been added to display and hide the modal
dialog for the portfolio section. The closeModal function waits for the
user to click the x in the dialog and will then hide it.

The openModal function takes the id attribute of the selector to save
as a variable and parse out the elements for the dialog box from a JSON
file. This will allow for better performance by only loading the larger
images when a user clicks on the accompanying project.