Add local development files and settings
The requisite files for bootstrap, font awesome, and jquery were added
as local assets to enable local development when no internet connection
is present. The default values in the various indexes and layout files
was replace with a liquid tag defined in _config.yml & _config.dev.yml.
The default set in _config.yml for each tag is to use the CDN method
of fetching the file. When developing without an internet connection
the _config.dev.yml file can be added as an argument for jekyll serve
which will overwrite the HTML to use the local files instead.