Add default values for blogging
Several default values were added as well as functionality for creating
posts with Prose.io. Author and default excerpt-seperator were defined.
The values for Prose mimic the default values for YML front-matter
if the user is writing posts in a text editor. Items that should not
be changed are hidden from the user with this posting method.