Fix numerous styling issues for blog index and post
An issue was discovered when including images in the post and how they
would line up with the remaining elements. Images needed to be styled
differently for the blog index and the post. In order to accomplish
this seperate stylesheets were created for both layouts. The layout
files were updated to reference the respective css files. Due to
nesting CSS declarations the user does not need to do any kramdown
style declarations in the markdown of each post.

Additional fixes:

* The size of the H1 elements is reduced slightly at the screen break
variable of 600px.
* The excerpt separator defaults in config.yml were corrected from
a previous spelling error.
* The pagination links for the blog index have been temporarily
commented out. When more than 5 posts exist the comment can be
removed. With only one page they cluttered the interface.