Implement Javascript promise to stage functions
A Javascript promise was needed to resolve when the quoteJSON variable
was fully poipulated. On resolve it initializes the user ability to
click the Get New Quote button to retrieve a new quote via the
masterUpdate function. Additionally, the resolve function calls the
first iteration of the masterUpdate function so a quote is displayed
for the user.