Reroute portfolio to projects subdirectory
Portfolio projects are now located under the /projects/ subdirectory
including the domain. This change was necessary to better automate
pulling in changes for each project via Gulp. Links in the JSON file
and portfolio.html will updated to reflect the new paths. Additionally,
the paths for different elements in each project were shifted to
relative paths to maintain functionality in their separate repos and
the main site when integrated.