Add site image assets to gulp imagemin task
The sites main assets/img folder has been included in the gulp imagemin
task. Additionally, the changed plugin was added to the workflow for
both this folder and the blog assets folder.

In order to complete this, the assets/img directory had  src and dist
subdirectories added to it. The image links in the site were updated to
point to the dist path.