Add resetIntervalTimer method to Carousel
This method is called to reset the timer to transition between items in
the carousel if a user clicks an arrow when `isPlaying` is true. The
logic was refactored in the updateProject method to accomodate this
fork. The left and right arrows now also pass in a second argument of
true when they are clicked. The interval call of the function omits
this argument which is defaulted to false. Tests were updated and also
written for the resetIntervalTimer method.