Create button component
The button component currently accepts three props. If it is passed an
`href` prop it will render out an anchor element link. If it is passed
a `to` prop it will render out a React link. If it is passed a button
`type` it will change the type of button. Currently, the button can
only have the types of 'button' and 'submit'.