Use custom React Scroll Link component
The current implementation of the React Scroll element does not include
an href property on the resulting rendered anchor tag. This affects the
accessibility features of tabbing through links and using enter to
activate them. In order to circumvent this, a custom implementation was
added that adds an href property as if targeting an id in the document.
This enables the accessibility features