Convert to class Component and add window scrolling
The component was converted to a class based component to add the
didMount life hook. This looks for a hash in the url location and
parses out the location of the window depending on the existence of
a hash. It also uses parent refs to target elements in child components