Refactor to eliminate Gatsby build errors
Several of the components included reference to the DOM window or
attempted to create new audio. This created build errors with Gatsby
which utilizes a SSR approach and doe not have a window to insert them
into. Additionally, a custom webpack config for Hammerjs was added to
not process it until after the build.